Align file extensions with documentation
The user documentation at https://eeditiones.github.io/tuttle-doc/getstarted/configuration/ states:

> Tuttle is configured in modules/config.xqm in your Tuttle installation.

This changes modules/config.xql to conform with the documentation’s filename.

In the process, it applies the community file extensions to all XQuery modules in the repository - namely, .xq for main modules, and .xqm for library modules.

It adjusts module prologs to use the updated location hints for module imports.

Also, for consistency, it adds newlines to modules that lack them.