fix(app): use standard shortened commit hash
Default git tools will produce a shortened hash of a commit to be
the first _seven_ characters.
Tuttle used the first _six_ instead which lead to those standard commit
hashes not being recognized.

BREAKING CHANGE: Because the commit hash was shortened to the first 6 characters
before, it is likely that comparisons against old and new short hashes will not succeed.
In order to update to this version of tuttle do a full clone after installing this version.

fixes #11