fix(github): handle edge-case when a file is modified and then renamed
That would have left the old file in the connected collection dangling.
Now every renamed file will be deleted.
As another edge-case where a file is _added_ and then renamed would try to remove an non-existing file, `github:incremental-delete`
will now treat a not found resource as a successful deletion.