fix(github): incremental update reports ignored files
Ignored files are expath-pkg.xml, repo.xml and all .xconf files.
Files that would be ignored was already reported by the dry-run but not for the actual run.
This allows for clients to act on that information and can, for instance, report that a full installation is necessary or that the index configuration has changed.